+++ pkgs/cronet_http/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.1-wip
 
+* Use `package:http_image_provider` in the example application.
 * Support Android API 21+.
 
 ## 1.0.0